+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - lux-file-list: Die onClick-Methode der `luxUploadActionConfig` wird bei der Löschaktion im Header pro Datei aufgerufen.
 - LuxMasterHeaderComponent.luxClicked -> LuxMasterHeaderComponent.luxOpened
 - LuxActionColorType -> LuxThemePalette
+- luxSelectedFiles -> luxSelected
 
 ### Defaultänderungen